refactor: domain/ 하위로, UI 관련 로직은 view/ 하위에서 관리
