+++ README.md
@@ -29,7 +29,7 @@
 
 ### 예외
 
-- [ ] 사용자가 잘못된 입력 값을 작성한 경우 에러 메시지를 보여주고, 다시 입력할 수 있게 한다.
+- [x] 사용자가 잘못된 입력 값을 작성한 경우 에러 메시지를 보여주고, 다시 입력할 수 있게 한다.
 - [x] 자동차 이름은 1자 이상 5자 이하이다.
 - [x] 자동차 이름에 공백이 들어가지 않는다.
 - [x] 자동차 이름이 중복되지 않는다.
@@ -39,7 +39,7 @@
 
 ## 입력 & 출력
 
-- 콤마를 기준으로 자동차 이름을 입력받는다. - UI
+- [x] 콤마를 기준으로 자동차 이름을 입력받는다. - UI
 - 횟수를 입력받는다. - UI
 - 전진하는 자동차의 이름을 출력한다 - UI
 - 전진하는 자동차의 상태를 출력한다 - UI
