+++ src/domain/model/CarManager.js
@@ -1,8 +1,8 @@
 // @ts-check
 import Car from './Car.js';
-import pickRandomNumber from '../utils/pickRandomNumber.js';
-import OutputView from '../views/OutputView.js';
-import { CONFIG } from '../constants/config.js';
+import pickRandomNumber from '../../utils/pickRandomNumber.js';
+import OutputView from '../../views/OutputView.js';
+import { CONFIG } from '../../constants/config.js';
 
 class CarManager {
   constructor() {
